+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - (Maintenance) Make scale_down_candidate annotation obsolete
 - (Bugfix) Fix ResignJob ID propagation
 - (Bugfix) Allow shards with RF1 in EnforcedResignLeadership action
+- (Maintenance) Update go-driver to v1.6.0, update IsNotFound() checks
  
 ## [1.2.33](https://github.com/arangodb/kube-arangodb/tree/1.2.33) (2023-09-27)
 - (Maintenance) Bump golang.org/x/net to v0.13.0